fix bug

+++ app/routes/index_routes.py
@@ -35,7 +35,8 @@ def index():
                     f"Reading video input parameters ({source_type=}, {upload.filename=})"
                 )
                 save_dir = os.path.join(current_app.root_path, "saved")
-                shutil.rmtree(save_dir)
+                if os.path.exists(save_dir):
+                    shutil.rmtree(save_dir)
                 os.makedirs(save_dir, exist_ok=True)
                 path = os.path.join(save_dir, upload.filename)
                 upload.save(path)
